requirements.txt
setuptools==67.2.0
flask-bootstrap == 3.3.7.1
Flask-RESTful == 0.3.9
SQLAlchemy == 2.0.4
mysql-connector-python == 8.0.32
python-dotenv == 1.0.0
flask-sqlalchemy == 3.0.3
flask-migrate == 4.0.4
pydantic==1.10.6